[apache] Remove all `php` configurations
The php directives have been included in the Apache hypertext access
file more as helpful information then to be used out of the box.

Nowadays, the use of the directives is highly dependent on the php
version and the way in witch php is integrated with Apache. This makes
them less usefull by default and just extra code to remove for both
the people that don't use php and the ones that do (especially a newer
version).

Reference: h5bp/server-configs#140