Add configs for `application/x-javascript`
By default the configurations from this repository ensure that
`.js` files are served with the `application/javascript` media
type, compressed, and with far future expires headers. However,
not all users¹ use all the configurations provided by this
project, or in some cases, the headers are overwritten from the
level "beneath" Apache (e.g.: from `php`).

Looking at the results from the HTTP Archive² from 01.09.2014:

  +---+--------------------+--------------------------+
  |   | number of requests |        media type        |
  +---+--------------------+--------------------------+
  | 1 |            7406888 | image/jpeg               |
  | 2 |            4547461 | image/png                |
  | 3 |            4003685 | image/gif                |
  | 4 |            3087744 | text/html                |
* | 5 |            2112777 | text/javascript          |
  | 6 |            1798370 | text/css                 |
* | 7 |            1740983 | application/x-javascript |
* | 8 |            1394717 | application/javascript   |
  |   |                ... |                      ... |
  +---+--------------------+--------------------------+

it's clear that the `application/javascript` media type is used
less often than both `text/javascript` and `application/x-javascript`.

Because of the above, the changes from this commit will ensure that
Apache will serve anything labeled with the `application/x-javascript`,
compressed and with far future expires headers (a previous commit³ did
the same for the files labeled as `text/javascript`).

- - - - - - - - - - - - - - - - - - - - - - - -- - - - - - - - - - - -

¹ h5bp/server-configs-apache#27
  h5bp/server-configs-apache#41

² https://www.igvita.com/2013/06/20/http-archive-bigquery-web-performance-answers/

  Query used:

    select count(requestid) as number_of_requests,
           mimetype as media_type
    from [httparchive:runs.2014_09_01_requests]
    group by media_type
    order by number_of_requests desc;

³ https://github.com/h5bp/server-configs-apache/commit/db6932740a90a36cbbf8b38627fc034d595471c0