Change media types for `.atom` and `.rss` files
Modify the configurations so that:

 * `.atom` files are served with the `application/atom+xml`¹ media type
 * `.rss` files are served with the `application/rss+xml`² media type

------------------------------------------------------------------------

Until now the configurations ensured that `.atom` and `.rss` files
were served with the `application/xml` media type, and that was done
in order to prevent³ some browsers from opening a download prompt
when users viewed these types of files directly.

That however, in combination with the `X-Content-Type-Options:
"nosniff"` header, prevents⁴ some browsers from doing something other
than just showing the raw content of the files.

By changing the media types we allow browsers to interpret the content
of the files and do more⁴ then just display them in raw form (e.g.:
show a nicer preview of the file, open the file in a local application),
and thus, we improve the user experience.

Note: One issue with this change is that IE 6 users will get the
download prompt, but that isn't a big problem nowadays.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

¹ https://www.iana.org/assignments/media-types/application/atom+xml
² http://tools.ietf.org/id/draft-nottingham-rss-media-type-00.txt
³ https://github.com/h5bp/html5-boilerplate/issues/989
⁴ https://github.com/h5bp/server-configs-apache/pull/50#issuecomment-76552855

------------------------------------------------------------------------

Thanks Gennadiy Litvinyuk (@gennadiylitvinyuk)!

Ref h5bp/server-configs-apache#50
Close h5bp/server-configs-apache#50