Update x-powered-by.conf
as per #183, `always` can be used to ensure the directive is always applied. In this case should be a sensed default. Also as found [here](https://www.tunetheweb.com/security/http-security-headers/server-header/) a good default is to use both versions since the `always` option is reported as not always working in some server configurations.