Refactor desc., highlight caution for disabling
I added that we're using `cache-control: no-transform` in the introductory paragraph because that is what we link to below that paragraph. Can remove if redundant, but added for clarity.

Also, IMO we should still separate the warnings "(!)" about considering the impact of disabling content transformations and "If you are using `mod_pagespeed`" because the majority of developers probably aren't using `mod_pagespeed` and so if you aren't using it, it's simple to distinguish.